fix load commands loop on arch_64 (otool)

+++ includes/otool.h
@@ -6,7 +6,7 @@
 /*   By: fkoehler <fkoehler@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/23 14:45:58 by fkoehler          #+#    #+#             */
-/*   Updated: 2017/11/23 19:44:26 by fkoehler         ###   ########.fr       */
+/*   Updated: 2017/11/29 12:01:37 by fkoehler         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ int					handle_lib(t_otool *env);
 
 int					init_env_struct(t_otool *env);
 
-void				print_section_64(char *sec_name, uint32_t sec_addr,
+void				print_section_64(char *sec_name, uint64_t sec_addr,
 					uint32_t sec_size, unsigned char *content);
 
 #endif
